+++ app/controllers/users_controller.rb
@@ -7,24 +7,13 @@ class UsersController < ApplicationController
 
   def create
     @user = User.new(user_params)
-    if @user.valid?
-      charge = StripeWrapper::Charge.create(
-          amount: 999, # amount in cents, again
-          source: params[:stripeToken],
-          description: "Sign up charge for #{@user.email}"
-        )
-      if charge.successful? 
-        @user.save
-        handle_invitation
-        AppMailer.delay.send_welcome_email(@user)
-        flash[:success] = "Thank you for registering with MyFlix. Please Sign in now."
-        redirect_to sign_in_path
-      else
-        flash.now[:error] = charge.error_message
-        render :new
-      end
+    result = UserSignup.new(@user).sign_up(params[:stripeToken], params[:invitation_token])
+
+    if result.successful?
+      flash[:success] = "Thank you for registering with MyFlix. Please Sign in now."
+      redirect_to sign_in_path
     else
-      flash.now[:error] = "Invalid user information. Please check the errors below."
+      flash[:error] = result.error_message
       render :new
     end
   end
@@ -45,21 +34,8 @@ class UsersController < ApplicationController
   end
 
   private
-    
+
   def user_params
     params.require(:user).permit(:email, :password, :full_name)
-  end
-  
-  def handle_invitation
-    if params[:invitation_token].present?
-      invitation = Invitation.find_by(token: params[:invitation_token])
-      @user.follow(invitation.inviter)
-      invitation.inviter.follow(@user)
-      remove_token(invitation, :token)
-    end
-  end
-  
-  def remove_token(invitation, token)
-      invitation.update_column(token, nil)
   end
 end
