+++ spec/cassettes/User_Registers/with_valid_user_info_and_invalid_card.yml
@@ -1,0 +1,124 @@
+---
+http_interactions:
+- request:
+    method: post
+    uri: https://api.stripe.com/v1/charges
+    body:
+      encoding: US-ASCII
+      string: amount=999&currency=usd&description=Sign+up+charge+for+john%40example.com&source=tok_17j0aeFbpquSrt2N1ELJVSBX
+    headers:
+      Accept:
+      - "*/*; q=0.5, application/xml"
+      Accept-Encoding:
+      - gzip, deflate
+      User-Agent:
+      - Stripe/v1 RubyBindings/1.31.0
+      Authorization:
+      - Bearer sk_test_C0Q8gFGxNZsXIpEvA4lsJJ7r
+      Content-Type:
+      - application/x-www-form-urlencoded
+      X-Stripe-Client-User-Agent:
+      - '{"bindings_version":"1.31.0","lang":"ruby","lang_version":"2.2.3 p173 (2015-08-18)","platform":"i686-linux","engine":"ruby","publisher":"stripe","uname":"Linux
+        version 3.13.0-39-generic (buildd@roseapple) (gcc version 4.8.2 (Ubuntu 4.8.2-19ubuntu1)
+        ) #66-Ubuntu SMP Tue Oct 28 13:31:23 UTC 2014","hostname":"admin008"}'
+      Content-Length:
+      - '109'
+  response:
+    status:
+      code: 200
+      message: OK
+    headers:
+      Server:
+      - nginx
+      Date:
+      - Sat, 27 Feb 2016 05:26:07 GMT
+      Content-Type:
+      - application/json
+      Content-Length:
+      - '1500'
+      Connection:
+      - keep-alive
+      Access-Control-Allow-Credentials:
+      - 'true'
+      Access-Control-Allow-Methods:
+      - GET, POST, HEAD, OPTIONS, DELETE
+      Access-Control-Allow-Origin:
+      - "*"
+      Access-Control-Max-Age:
+      - '300'
+      Cache-Control:
+      - no-cache, no-store
+      Request-Id:
+      - req_7yyXrmgCtcUGH0
+      Stripe-Version:
+      - '2015-10-16'
+      Strict-Transport-Security:
+      - max-age=31556926; includeSubDomains
+    body:
+      encoding: UTF-8
+      string: |
+        {
+          "id": "ch_17j0ahFbpquSrt2NldLXIooC",
+          "object": "charge",
+          "amount": 999,
+          "amount_refunded": 0,
+          "application_fee": null,
+          "balance_transaction": "txn_17j0ahFbpquSrt2NpqufLpli",
+          "captured": true,
+          "created": 1456550767,
+          "currency": "usd",
+          "customer": null,
+          "description": "Sign up charge for john@example.com",
+          "destination": null,
+          "dispute": null,
+          "failure_code": null,
+          "failure_message": null,
+          "fraud_details": {},
+          "invoice": null,
+          "livemode": false,
+          "metadata": {},
+          "order": null,
+          "paid": true,
+          "receipt_email": null,
+          "receipt_number": null,
+          "refunded": false,
+          "refunds": {
+            "object": "list",
+            "data": [],
+            "has_more": false,
+            "total_count": 0,
+            "url": "/v1/charges/ch_17j0ahFbpquSrt2NldLXIooC/refunds"
+          },
+          "shipping": null,
+          "source": {
+            "id": "card_17j0aeFbpquSrt2NNtdRleH2",
+            "object": "card",
+            "address_city": null,
+            "address_country": null,
+            "address_line1": null,
+            "address_line1_check": null,
+            "address_line2": null,
+            "address_state": null,
+            "address_zip": null,
+            "address_zip_check": null,
+            "brand": "Visa",
+            "country": "US",
+            "customer": null,
+            "cvc_check": "pass",
+            "dynamic_last4": null,
+            "exp_month": 7,
+            "exp_year": 2018,
+            "fingerprint": "P2iXG4w2kPZzpxZL",
+            "funding": "credit",
+            "last4": "4242",
+            "metadata": {},
+            "name": null,
+            "tokenization_method": null
+          },
+          "source_transfer": null,
+          "statement_descriptor": null,
+          "status": "succeeded"
+        }
+    http_version: 
+  recorded_at: Sat, 27 Feb 2016 05:26:07 GMT
+recorded_with: VCR 3.0.1
