+++ spec/features/user_registers_spec.rb
@@ -7,6 +7,7 @@ feature 'User Registers', :js, :vcr do
     fill_in_valid_user_info
     fill_in_valid_card
     click_button "Sign Up"
+    sleep 0.2
     expect(page).to have_content("Thank you for registering with MyFlix. Please Sign in now.")
   end
 
@@ -14,12 +15,14 @@ feature 'User Registers', :js, :vcr do
     fill_in_valid_user_info
     fill_in_invalid_card
     click_button "Sign Up"
-    expect(page).to have_content("This card number looks invalid.")
+    sleep 0.2 
+    expect(page).to have_content("The card number is not a valid credit card number.")
   end
   scenario "with valid user info and declined card" do
     fill_in_valid_user_info
     fill_in_declined_card
     click_button "Sign Up"
+    sleep 0.25
     expect(page).to have_content("Your card was declined.")
   end
 
@@ -27,6 +30,7 @@ feature 'User Registers', :js, :vcr do
     fill_in_invalid_user_info
     fill_in_valid_card
     click_button "Sign Up"
+    sleep 0.2
     expect(page).to have_content("Invalid user information. Please check the errors below.")
   end
 
@@ -34,14 +38,16 @@ feature 'User Registers', :js, :vcr do
     fill_in_invalid_user_info
     fill_in_invalid_card
     click_button "Sign Up"
-    expect(page).to have_content("This card number looks invalid.")
+    sleep 0.25
+    expect(page).to have_content("The card number is not a valid credit card number.")
   end
   
   scenario "with invalid user info and declined card" do
     fill_in_invalid_user_info
     fill_in_declined_card
     click_button "Sign Up"
-    expect(page).to have_content("Invalid user information. Please check the errors below.")
+    sleep 0.2
+    expect(page).to have_content("Please fix the error below")
   end
 
   def fill_in_valid_user_info
